Refactor types.js to common.js
