+++ lib/Pill/PillBase.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { types, defaults } from './pillTypes';
+import { types, defaults } from './common';
 
 const PillBase = ({ className, children, ...rest }) => (
   <span
