+++ lib/Pill/PillGreen.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { types, defaults } from './pillTypes';
+import { types, defaults } from './common';
 import PillBase from './PillBase';
 
 const PillGreen = ({ className, children, ...rest }) => (
