+++ lib/Pill/PillRed.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { types, defaults } from './pillTypes';
+import { types, defaults } from './common';
 import PillBase from './PillBase';
 
 const PillRed = ({ className, children, ...rest }) => (
