+++ lib/Pill/common.js
@@ -1,5 +1,10 @@
 import { node, string } from 'prop-types';
 
+export const sizes = {
+  md: 'md',
+  xs: 'xs'
+};
+
 export const types = {
   children: node.isRequired,
   className: string
