+++ lib/Pill/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { string } from 'prop-types';
-import { types, defaults } from './pillTypes';
+import { types, defaults } from './common';
 import PillDefault from './PillDefault';
 import PillRed from './PillRed';
 import PillGreen from './PillGreen';
